+++ AbilityScripts/Extra/SpawnAbilityInfo.cs
@@ -1,4 +1,5 @@
 ﻿using GameNetcodeStuff;
+using HideAndSeek.Patches;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -8,5 +9,27 @@ namespace HideAndSeek.AbilityScripts.Extra
     {
         public PlayerControllerB creatorPlayer;
         public List<PlayerControllerB> otherFriendlies = new();
+
+        void Start()
+        {
+            PatchesManager.playerRevived += PlayerRevived;
+        }
+
+        void PlayerRevived(ulong id)
+        {
+            PlayerControllerB newZombie = RoundManagerPatch.GetPlayerWithClientId(id);
+            if (Plugin.seekers.Contains(creatorPlayer))
+            {
+                // Creator is seeker
+                if(!otherFriendlies.Contains(newZombie))
+                    otherFriendlies.Add(newZombie);
+            }
+            else
+            {
+                // Creator is hider
+                if (otherFriendlies.Contains(newZombie))
+                    otherFriendlies.Remove(newZombie);
+            }
+        }
     }
 }